Added default implementation for interrupt traits
In the InterruptSourceGroup trait the only method that is going to
be implemented for every interrupt type is the trigger method. We
can make it easier for simple VMMs (like the ones using only legacy
IRQs) to implement this trait by provided no-op implementation for
all the other trait methods.

Signed-off-by: Andreea Florescu <fandree@amazon.com>